requirements.txt
Django==2.2.5
django-mdeditor==0.1.16
Pillow==6.0.0
simplejson==3.16.0
djangorestframework==3.10.3
channels==2.3.0
channels-redis==2.4.0
stripe==2.37.2
psycopg2-binary==2.8.3
psycopg2==2.8.3
pylint-django==2.0.11